# Conflicts:
#	.idea/design-patterns-final-project.iml
#	.idea/inspectionProfiles/Project_Default.xml
#	.idea/misc.xml

+++ core/repository_interface/wallet_repository_interface.py
@@ -2,16 +2,16 @@ from typing import Any, Optional, Protocol
 
 
 class IWalletRepository(Protocol):
-    def create_wallet(self, user: str, address: str, btc_balance: int = 1) -> bool:
+    def create_wallet(self, user: str, address: str, btc_balance: float = 1) -> bool:
         pass
 
     def exists_wallet(self, address: str) -> bool:
         pass
 
-    def change_balance(self, address: str, balance_change: int) -> None:
+    def change_balance(self, address: str, balance_change: float) -> None:
         pass
 
-    def get_balance(self, address: str) -> int:
+    def get_balance(self, address: str) -> float:
         pass
 
     def get_user(self, address: str) -> str:
